Auto-commit after MAX_CHANGES modified entries
This used to be a feature of `automatic-corrections.py` and has now been
moved to the middleware framework. This way we ensure that automatic
bulk edits via the framework will always be split into conveniently sized
commits of (currently) max. 100 automatically changed files (plus
possibly a few files that were altered manually at the same time).